requirements.txt
Pillow>=9.5.0
pytesseract>=0.3.10
ttkbootstrap>=1.10.1
tkinterdnd2>=0.3.0